requirements.txt
aiogram==3.*
aiosqlite==0.*
google-api-python-client==2.*
google-auth==2.*
apscheduler==3.*
icalendar==6.*
pydantic-settings==2.*
python-dotenv==1.*
